// src/app/models/sales.model.ts
export interface Sale {
    id: number;
    transactionCode: string;
    storeCode: string;
    productCode: any;
    productName: string;
    quantity: string;
    storeQuantity: any;
    date: string;
    discountPercent: string;
    discountAmount: string;
    discountNote: string;
    amount: string;
    firstName: string;
    lastName: string;
    address: string;
    phone: string;
    status: string;
    image: any;
}